server/requirements.txt
flask
flask-cors
pillow
gunicorn
--extra-index-url https://download.pytorch.org/whl/cpu
torch
ultralytics
